Fixed signing root test
